Prepare RDS data for Bootstrap Re-sampling Following Salganik (2006)

Prepare RDS data for Bootstrap Re-sampling Following Salganik (2006)

get_rds_boot(
  data,
  seed_condition = "rds_from == -999",
  in_coupon = "rds_own_coupon",
  out_coupon = "rds_coupon_",
  is_child_fn = is_child,
  trait_var = "hidden_visible_out",
  other_vars = c("hidden_visible_out", "known_visible_out", "hidden"),
  n_boot = 100
)

Arguments

data

sample data frame

seed_condition

character string containing condition to define seeds. Defaults to "rds_from == -999" that applies to simulated RDS samples

in_coupon

character string for variable containing in coupons identifiers. In coupon is coupon used to enroll specific respondent

out_coupon

character string for variable(s) containing out coupons identifiers. Out coupons are coupons given to respondent to enroll other study participants from their network

is_child_fn

function used to identify direct childs of specific seed. Defaults to function that uses "-" in the coupon identifiers

trait_var

variable used to split RDS sample for resampling. Defaults to "hidden_visible_out" that captures number of hidden population members recognized by respondent

other_vars

other variables to preserve in bootstrap samples

n_boot

number of bootstrap resamples

References

Dennis M. Feehan, Matthew J. Salganik. “The surveybootstrap package.” (2016). https://cran.r-project.org/package=surveybootstrap.