Modified Function for Rescaled Bootstrap by Rust and Rao (1996)

Modified Function for Rescaled Bootstrap by Rust and Rao (1996)

get_rescaled_boot(data, survey_design, n_boot = 10)

Arguments

data
sample data frame

survey_design
a formula describing the design of the survey

n_boot
number of bootstrap resamples

Value

References

Dennis M. Feehan, Matthew J. Salganik. “The surveybootstrap package.” (2016). https://cran.r-project.org/package=surveybootstrap. Rust, Keith F., and J. N. K. Rao. "Variance estimation for complex surveys using replication techniques." Statistical methods in medical research 5, no. 3 (1996): 283-310.

Examples