Create Diagnosis Plot of Log Normalized RMSE Across Studies

Create Diagnosis Plot of Log Normalized RMSE Across Studies

get_rmse_plots(
  diagnosands_df,
  study_map = c(`Brazil (FF)` = "ff_brazil", `Brazil (Stanford)` = "stanford_brazil",
    `Costa Rica (John Jay)` = "johnjay_costarica", `Morocco (NORC)` = "norc_marocco",
    `Pakistan (JHU)` = "jhu_pakistan", `Tanzania (John Jay)` = "johnjay_tanzania",
    `Tunisia (UMass)` = "umass_tunisia", `USA (RTI)` = "rti_usa"),
  sampling_map = c(PPS = "pps", RDS = "rds", TLS = "tls", `RDS/TLS` = "rds_tls",
    `RDS/PPS` = "rds_pps"),
  estimator_map = c(NSUM = "nsum", HT = "ht", Recapture = "recap", `SS-PSE` = "sspse",
    `Service Multiplier` = "multi", Chords = "chords")
)

Arguments

diagnosands_df
diagnosands data frame produced by diagnose_design()

study_map
named character vector of mapping between study labels and names used in the plot

sampling_map
named character vector of mapping between sampling strategies and names used in the plot

estimator_map
named character vector of mapping between estimators and names used in the plot

Value

Examples