Chords population size estimatior by Berchenko, Rosenblatt and Frost

Chords population size estimatior by Berchenko, Rosenblatt and Frost

get_study_est_chords(
  data,
  type = c("mle", "integrated", "jeffreys"),
  seed_condition = "rds_from == -999",
  rds_prefix = "rds",
  n_boot = 100,
  parallel_boot = FALSE,
  label = "chords"
)

Arguments

data
pass-through population data frame

type
a character vector with the type of estimation. Can be one of mle, integrated, or jeffreys. See ?chords::Estimate.b.k and the original paper from the references for details

seed_condition
character string containing condition to define seeds. Defaults to "rds_from == -999" that applies to simulated RDS samples

rds_prefix
character prefix used for RDS sample variables

n_boot
number of bootstrap resamples

parallel_boot
logical, whether to compute bootstrap samples in parallel using foreach package

label
character string describing the estimator

Value

Data frame of Chords estimates for a single study with RDS sample

References

Berchenko, Yakir, Jonathan D. Rosenblatt, and Simon D. W. Frost. “Modeling and Analyzing Respondent-Driven Sampling as a Counting Process.” Biometrics 73, no. 4 (2017): 1189–98. https://doi.org/10.1111/biom.12678. Salganik, Matthew J. "Variance estimation, design effects, and sample size calculations for respondent-driven sampling." Journal of Urban Health 83, no. 1 (2006): 98. https://doi.org/10.1007/s11524-006-9106-x