Estimator of population size based on Link-Tracing Sample by Vincent and Thompson

Estimator of population size based on Link-Tracing Sample by Vincent and Thompson

get_study_est_linktrace(
  data,
  total = 2000,
  strata,
  gibbs_params = list(n_samples = 50L, chain_samples = 250L, chain_burnin = 50L),
  priors = list(p_n = 0L, p_l = 0.1, p_b = 1L),
  progress = FALSE,
  prefix = "lts",
  label = "link"
)

Arguments

data

pass through sample

total

integer giving the total size of the population

strata

string specifying column name of strata vector

gibbs_params

named list of parameters passed to Gibbs sampler

priors

named list of prior specification for population size, stratum membership and links. p_n is an integer specifying the power law prior for population size (0 = flat). p_l is a positive rational numeric vector of length n_strata specifying the dirichlet prior for stratum membership (0.1 = non-informative). p_b is an integer specifying the beta distribution prior for links (1 = non-informative).

progress

logical indicating whether to display progress bar. Defaults to FALSE

prefix

character string giving name of the column with RDS+ sampling indicator

label

character string giving label for the estimator

Value

Data frame of link tracing estimates for single study

References

Vincent, Kyle, and Steve Thompson. "Estimating population size with link-tracing sampling." Journal of the American Statistical Association 112.519 (2017): 1286-1295. Vincent, Kyle, and Steve Thompson. "Estimating the size and distribution of networked populations with snowball sampling." Journal of Survey Statistics and Methodology 10.2 (2022): 397-418.