NSUM estimatior

NSUM estimatior

get_study_est_nsum(
  data,
  known,
  hidden,
  survey_design = ~pps_cluster + strata(pps_strata),
  degree_ratio = 1,
  transmission_rate = 1,
  n_boot = 1000,
  parallel_boot = FALSE,
  prefix = "pps",
  label = "nsum"
)

Arguments

data
pass-through population data frame

known
character vector containing names of known groups

hidden
character vector containing names of hidden groups

survey_design
a formula describing the design of the survey

degree_ratio
numeric value between 0 and 1 representing degree ratio

transmission_rate
numeric value between 0 and 1 representing information transmission rate

n_boot
number of bootstrap resamples

parallel_boot
logical, whether to compute bootstrap samples in parallel using foreach package

prefix
character prefix used for PPS sample variables

label
character string describing the estimator

Value

Data frame of NSUM estimates for a single study with PPS sample

References

Dennis M. Feehan, Matthew J. Salganik. “The networkreporting package.” (2014). https://cran.r-project.org/package=networkreporting. Dennis M. Feehan, Matthew J. Salganik. “The surveybootstrap package.” (2016). https://cran.r-project.org/package=surveybootstrap. Salganik, Matthew J. "Variance estimation, design effects, and sample size calculations for respondent-driven sampling." Journal of Urban Health 83, no. 1 (2006): 98. https://doi.org/10.1007/s11524-006-9106-x