Sequential Sampling (SS) prevalence estimator by Gile (2011)

Sequential Sampling (SS) prevalence estimator by Gile (2011)

get_study_est_rds_ss(
  data,
  hidden_var = "hidden",
  n_coupons = 3,
  total = 2000,
  rds_prefix = "rds",
  label = "sspse"
)

Arguments

data
pass-through population data frame

hidden_var
character string specifying hidden variable name (associated probability of visibility should be named p_visible_[hidden_var]). Defaults to "hidden" for the simulations

n_coupons
The number of recruitment coupons distributed to each enrolled subject (i.e. the maximum number of recruitees for any subject). By default it is taken by the attribute or data, else the maximum recorded number of coupons.

total
integer giving the total size of known population (denominator for prevalence)

rds_prefix
character prefix used for RDS sample variables

label
character string describing the estimator

Value

References

Gile, Krista J. 2011 Improved Inference for Respondent-Driven Sampling Data with Application to HIV Prevalence Estimation, Journal of the American Statistical Association, 106, 135-146. Gile, Krista J., Handcock, Mark S., 2010 Respondent-driven Sampling: An Assessment of Current Methodology, Sociological Methodology, 40, 285-327.