Link-tracing Gibbs sampler

Link-tracing Gibbs sampler

lt_gibbs_cpp(
  links_list,
  wave,
  name,
  y_samp,
  strata,
  n_strata,
  n_waves,
  total,
  chain_samples,
  chain_burnin,
  prior_n,
  prior_l,
  prior_b,
  n_0,
  l_0,
  b_0,
  n_samples,
  progress
)

Arguments

links_list

list of between unit edges

wave

integer vector of rds wave units were sampled in

name

integer vector of unit ids

y_samp

matrix pass through of adjecency matrix generated in get_study_est_linktrace

strata

integer stratum id of each unit

n_strata

integer number of unique strata

n_waves

integer number of sampling waves

total

integer total size of the population

chain_samples

integer number of samples per MCMC chain

prior_n

integer power law prior for population size

prior_l

double vector of dirichilet priors for stratum membership

prior_b

integer beta distribution prior for unit links

n_0

integer initial value for n

l_0

double vector initial values for l

b_0

double matrix initial values for b

n_samples

number of samples to draw

progress

bool indicating whether to display progress bar

chain_burning

integer number of burnin samples per MCMC chain

Value

a vector of vectors with n_samples population size samples